Refactor attempts history table to a separate component
