+++ frontend/src/components/Profile.jsx
@@ -1,4 +1,8 @@
+import {useEffect} from "react";
+import {getAttemptHistory} from "@app/api/attemptsService";
+
 const Profile = () => {
+
   return (
     <div>Profile</div>
   );
